:recycle: api: migrating CPF script as a usecase in session bounded context
