+++ api/scripts/certification/get-cpf-import-status-from-xml.js
@@ -1,14 +1,9 @@
-import fs from 'fs';
-import sax from 'sax';
-import saxPath from 'saxpath';
-import { logger } from '../../lib/infrastructure/logger.js';
-
-import xml2js from 'xml2js';
+import { logger } from '../../src/shared/infrastructure/utils/logger.js';
 
 import { disconnect } from '../../db/knex-database-connection.js';
-import * as cpfCertificationResultRepository from '../../src/certification/session/infrastructure/repositories/cpf-certification-result-repository.js';
-import { CpfImportStatus } from '../../src/certification/session/domain/models/CpfImportStatus.js';
 import * as url from 'url';
+
+import { getCPfImportstatusFromXml } from '../../src/certification/session/domain/usecases/get-cpf-import-status-from-xml.js';
 
 const modulePath = url.fileURLToPath(import.meta.url);
 const isLaunchedFromCommandLine = process.argv[1] === modulePath;
@@ -17,24 +12,8 @@ async function main(path) {
   logger.info("Récupération des résultats d'import CPF...");
   const filePath = path || process.argv[2];
 
-  const { cpfImportErrorIds, cpfImportSuccessIds } = await getCpfImportResults(filePath);
+  await getCPfImportstatusFromXml({ filePath });
 
-  if (cpfImportErrorIds.length) {
-    await cpfCertificationResultRepository.updateCertificationImportStatus({
-      certificationCourseIds: cpfImportErrorIds,
-      cpfImportStatus: CpfImportStatus.ERROR,
-    });
-  }
-
-  if (cpfImportSuccessIds.length) {
-    await cpfCertificationResultRepository.updateCertificationImportStatus({
-      certificationCourseIds: cpfImportSuccessIds,
-      cpfImportStatus: CpfImportStatus.SUCCESS,
-    });
-  }
-
-  logger.info(`${cpfImportErrorIds.length} certifications en erreur`);
-  logger.info(`${cpfImportSuccessIds.length} certifications importées avec succès`);
   logger.info(`Done! ...`);
 }
 
@@ -51,45 +30,4 @@ async function main(path) {
   }
 })();
 
-function getCpfImportResults(filePath) {
-  const xmlStream = fs.createReadStream(filePath);
-  const saxStream = sax.createStream({
-    strict: true,
-  });
-  return new Promise((resolve, reject) => {
-    const cpfImportErrors = new saxPath.SaXPath(saxStream, '//passage');
-    const cpfImportSuccess = new saxPath.SaXPath(saxStream, '//PassagesOK/idTechnique');
-
-    const cpfImportErrorIds = [];
-    const cpfImportSuccessIds = [];
-
-    cpfImportErrors.on('match', (xmlNode) => {
-      xml2js.parseString(xmlNode, async (err, { passage }) => {
-        const [errorMessage] = passage['messageErreur'];
-        const [certificationId] = passage.idTechnique;
-
-        if (errorMessage.includes('existe déjà en base')) {
-          cpfImportSuccessIds.push(certificationId);
-        }
-        if (errorMessage.includes('Aucun titulaire')) {
-          cpfImportErrorIds.push(certificationId);
-        }
-      });
-    });
-
-    cpfImportSuccess.on('match', (xmlNode) => {
-      xml2js.parseString(xmlNode, async (err, { idTechnique }) => {
-        cpfImportSuccessIds.push(idTechnique);
-      });
-    });
-
-    xmlStream
-      .pipe(saxStream)
-      .on('error', reject)
-      .on('end', async () => {
-        resolve({ cpfImportErrorIds, cpfImportSuccessIds });
-      });
-  });
-}
-
-export { getCpfImportResults, main };
+export { main };
