+++ api/tests/certification/session/unit/domain/usecases/files/xml/cpfImportLogEmpty.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<rapport xmlns:ns2="urn:cdc:cpf:pc5:rapport:schema">
+  <nomFichier>pix-cpf-export-20221003-324234.xml</nomFichier>
+  <idFlux>1234-abcd</idFlux>
+  <idTraitement>1</idTraitement>
+</rapport>
